LSP : EnvelopeGenerator - エンベロープモデルをメロディとドラムで明確に分離

+++ libsynthpp/src/LSP/Synth/Voice.cpp
@@ -69,12 +69,12 @@ void Voice::setPitchBend(float pitchBend)noexcept
 void Voice::setCutOff(float freqRate, float cutOffGain)
 {
 	float freq = mCalculatedFreq * freqRate;
-	mCutOffFilter.setHighshelfParam(static_cast<float>(mSampleFreq), freq, 1.f, cutOffGain);
+	mCutOffFilter.setHighshelfParam(mSampleFreq, freq, 1.f, cutOffGain);
 }
 void Voice::setResonance(float freqRate, float overtoneGain)
 {
 	float freq = mCalculatedFreq * freqRate;
-	mResonanceFilter.setPeakingParam(static_cast<float>(mSampleFreq), freq, 1.f, overtoneGain);
+	mResonanceFilter.setPeakingParam(mSampleFreq, freq, 1.f, overtoneGain);
 }
 Voice::EnvelopeGenerator& Voice::envolopeGenerator() noexcept
 {
